requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
PyPDF2==3.0.1
ebooklib==0.18
python-docx==1.1.0
beautifulsoup4==4.12.3
striprtf==0.0.26
mobi==0.3.3
odfpy==1.4.1
langdetect==1.0.9
lxml==5.1.0
